faceUI/requirements.txt
flask==3.0.0
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0